# docs/audit/2026-05-05-api-openapi.yaml
openapi: 3.0.3
info:
  title: PaletyBroker API
  version: 0.1.0-audit
  description: |
    Specyfikacja odzwierciedla aktualna implementacje endpointow.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:4000
paths:
...
  /admin/stats:
    get:
      summary: Statystyki panelu administracyjnego
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats payload
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (admin role required)
  /admin/pricing-rules:
    get:
      summary: Lista reguł cenowych (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pricing rules list
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (admin role required)
    post:
      summary: Utworzenie/aktualizacja reguły cenowej (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [carrierCode, serviceName, basePrice]
              properties:
                id: { type: string }
                carrierCode: { type: string }
                serviceName: { type: string }
                basePrice: { type: string }
                kmRate: { type: string }
                marginPercent: { type: string }
                minWeight: { type: string }
                maxWeight: { type: string }
                isActive: { type: boolean }
      responses:
        "201":
          description: Rule created or updated
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (admin role required)
...